* CMXCPHits_v1, CMXJetHits_v1: change size of hits0 and hits1 from uint8_t to uint32_t   (it happens that these fields can store extra information not only thresholds) * CPMTower_v2, CMXCPTob_v1, CMXEtSums_v1, JetElement_v2, CMXJetTob_v1, CMXCPHits_v1:   - change errors size from uint8_t to uint32_t (for unified usage in our tools,      this fields can store also additional bits from error status words) (xAODTrigL1Calo-00-00-55)
  * CMXCPHits_v1, CMXJetHits_v1: change size of hits0 and hits1 from uint8_t to uint32_t
    (it happens that these fields can store extra information not only thresholds)
  * CPMTower_v2, CMXCPTob_v1, CMXEtSums_v1, JetElement_v2, CMXJetTob_v1, CMXCPHits_v1:
    - change errors size from uint8_t to uint32_t (for unified usage in our tools,
       this fields can store also additional bits from error status words)
  * Tag as xAODTrigL1Calo-00-00-55

2015-08-04 Alexander Mazurov <alexander.mazurov@cern.ch>
  * CPMTobRoI_v1.cxx: add initialize method
  * Tag as xAODTrigL1Calo-00-00-54

2015-07-27 Alexander Mazurov <alexander.mazurov@cern.ch>
  * RODHeader and JEMTobRoI: add initialize methods.
  * Tag as xAODTrigL1Calo-00-00-53

2015-07-07 John Morris <john.morris@cern.ch>
	* CMXRoI
	* Tag as xAODTrigL1Calo-00-00-52

2015-07-02 John Morris <john.morris@cern.ch>
...
(Long ChangeLog diff - truncated)

+++ Event/xAOD/xAODTrigL1Calo/Root/CMXCPHits_v1.cxx
@@ -2,7 +2,7 @@
   Copyright (C) 2002-2017 CERN for the benefit of the ATLAS collaboration
 */
 
-// $Id: CMXCPHits_v1.cxx 663426 2015-04-28 11:45:23Z amazurov $
+// $Id: CMXCPHits_v1.cxx 687949 2015-08-06 15:48:49Z amazurov $
 
 // EDM includes(s):
 #include "xAODCore/AuxStoreAccessorMacros.h"
@@ -27,10 +27,10 @@ namespace xAOD{
   
   /// initialize
   void CMXCPHits_v1::initialize(const uint8_t crate,const uint8_t cmx,const uint8_t source,
-                                const std::vector<uint8_t>& hitsVec0,
-                                const std::vector<uint8_t>& hitsVec1,
-                                const std::vector<uint8_t>& errorVec0,
-                                const std::vector<uint8_t>& errorVec1,
+                                const std::vector<uint32_t>& hitsVec0,
+                                const std::vector<uint32_t>& hitsVec1,
+                                const std::vector<uint32_t>& errorVec0,
+                                const std::vector<uint32_t>& errorVec1,
                                 const uint8_t peak)
   {
     setCrate( crate );
@@ -44,10 +44,10 @@ namespace xAOD{
   } 
       
   /// add hits to existing object  
-  void CMXCPHits_v1::addHits(const std::vector<uint8_t>& hitsVec0,
-                             const std::vector<uint8_t>& hitsVec1,
-                             const std::vector<uint8_t>& errorVec0,
-                             const std::vector<uint8_t>& errorVec1)
+  void CMXCPHits_v1::addHits(const std::vector<uint32_t>& hitsVec0,
+                             const std::vector<uint32_t>& hitsVec1,
+                             const std::vector<uint32_t>& errorVec0,
+                             const std::vector<uint32_t>& errorVec1)
   {
     setHitsVec0( hitsVec0 );
     setHitsVec1( hitsVec1 );
@@ -59,35 +59,35 @@ namespace xAOD{
   AUXSTORE_PRIMITIVE_SETTER_AND_GETTER( CMXCPHits_v1 , uint8_t , cmx , setCmx )
   AUXSTORE_PRIMITIVE_SETTER_AND_GETTER( CMXCPHits_v1 , uint8_t , sourceComponent , setSourceComponent )
   AUXSTORE_PRIMITIVE_SETTER_AND_GETTER( CMXCPHits_v1 , uint8_t , peak , setPeak )
-  AUXSTORE_OBJECT_SETTER_AND_GETTER( CMXCPHits_v1 , std::vector<uint8_t> , hitsVec0 , setHitsVec0 )
-  AUXSTORE_OBJECT_SETTER_AND_GETTER( CMXCPHits_v1 , std::vector<uint8_t> , hitsVec1 , setHitsVec1 )
-  AUXSTORE_OBJECT_SETTER_AND_GETTER( CMXCPHits_v1 , std::vector<uint8_t> , errorVec0 , setErrorVec0 )
-  AUXSTORE_OBJECT_SETTER_AND_GETTER( CMXCPHits_v1 , std::vector<uint8_t> , errorVec1 , setErrorVec1 )
+  AUXSTORE_OBJECT_SETTER_AND_GETTER( CMXCPHits_v1 , std::vector<uint32_t> , hitsVec0 , setHitsVec0 )
+  AUXSTORE_OBJECT_SETTER_AND_GETTER( CMXCPHits_v1 , std::vector<uint32_t> , hitsVec1 , setHitsVec1 )
+  AUXSTORE_OBJECT_SETTER_AND_GETTER( CMXCPHits_v1 , std::vector<uint32_t> , errorVec0 , setErrorVec0 )
+  AUXSTORE_OBJECT_SETTER_AND_GETTER( CMXCPHits_v1 , std::vector<uint32_t> , errorVec1 , setErrorVec1 )
 
   /// Peak functions - return value of peak bunch crossing
   /// Assuming we have got the timing calibration correct,
   /// This is the bunch crossing containing the collision
 
   /// get hitsVec0 at peak bunch crossing
-  uint8_t CMXCPHits_v1::hits0() const
+  uint32_t CMXCPHits_v1::hits0() const
   {
     return hitsVec0()[ peak() ];
   }
 
   /// get hitsVec1 at peak bunch crossing
-  uint8_t CMXCPHits_v1::hits1() const
+  uint32_t CMXCPHits_v1::hits1() const
   {
     return hitsVec1()[ peak() ];
   }
 
   /// get errorVec0 at peak bunch crossing
-  uint8_t CMXCPHits_v1::error0() const
+  uint32_t CMXCPHits_v1::error0() const
   {
     return errorVec0()[ peak() ];
   }
 
   /// get errorVec1 at peak bunch crossing
-  uint8_t CMXCPHits_v1::error1() const
+  uint32_t CMXCPHits_v1::error1() const
   {
     return errorVec1()[ peak() ];
   }
